refactor: 점수 계산 GameResult가 하도록 수정
